+++ server/src/main/java/org/elasticsearch/search/aggregations/bucket/BucketsAggregator.java
@@ -107,11 +107,11 @@ public abstract class BucketsAggregator extends AggregatorBase {
      *
      * Refer to that method for documentation about the merge map.
      *
-     * @deprecated use {@link mergeBuckets(long, LongUnaryOperator)}
+     * @deprecated use {@link rewriteBuckets(long, LongUnaryOperator)}
      */
     @Deprecated
     public final void mergeBuckets(long[] mergeMap, long newNumBuckets) {
-        mergeBuckets(newNumBuckets, bucket -> mergeMap[Math.toIntExact(bucket)]);
+        rewriteBuckets(newNumBuckets, bucket -> mergeMap[Math.toIntExact(bucket)]);
     }
 
     /**
@@ -119,10 +119,10 @@ public abstract class BucketsAggregator extends AggregatorBase {
      *  @param mergeMap a unary operator which maps a bucket's ordinal to the ordinal it should be merged with.
      *  If a bucket's ordinal is mapped to -1 then the bucket is removed entirely.
      *
-     * This only tidies up doc counts. Call {@link MergingBucketsDeferringCollector#mergeBuckets(LongUnaryOperator)} to
+     * This only tidies up doc counts. Call {@link BestBucketsDeferringCollector#rewriteBuckets(LongUnaryOperator)} to
      * merge the actual ordinals and doc ID deltas.
      */
-    public final void mergeBuckets(long newNumBuckets, LongUnaryOperator mergeMap){
+    public final void rewriteBuckets(long newNumBuckets, LongUnaryOperator mergeMap){
         try (IntArray oldDocCounts = docCounts) {
             docCounts = bigArrays.newIntArray(newNumBuckets, true);
             docCounts.fill(0, newNumBuckets, 0);
