+++ server/src/main/java/org/elasticsearch/search/aggregations/bucket/DeferableBucketAggregator.java
@@ -37,7 +37,7 @@ public abstract class DeferableBucketAggregator extends BucketsAggregator {
      * Wrapper that records collections. Non-null if any aggregations have
      * been deferred.
      */
-    private DeferringBucketCollector recordingWrapper;
+    private DeferringBucketCollector deferringCollector;
     private List<String> deferredAggregationNames;
 
     protected DeferableBucketAggregator(String name, AggregatorFactories factories, SearchContext context, Aggregator parent,
@@ -52,28 +52,37 @@ public abstract class DeferableBucketAggregator extends BucketsAggregator {
         List<BucketCollector> deferredAggregations = null;
         for (int i = 0; i < subAggregators.length; ++i) {
             if (shouldDefer(subAggregators[i])) {
-                if (recordingWrapper == null) {
-                    recordingWrapper = getDeferringCollector();
+                if (deferringCollector == null) {
+                    deferringCollector = buildDeferringCollector();
                     deferredAggregations = new ArrayList<>(subAggregators.length);
                     deferredAggregationNames = new ArrayList<>(subAggregators.length);
                 }
                 deferredAggregations.add(subAggregators[i]);
                 deferredAggregationNames.add(subAggregators[i].name());
-                subAggregators[i] = recordingWrapper.wrap(subAggregators[i]);
+                subAggregators[i] = deferringCollector.wrap(subAggregators[i]);
             } else {
                 collectors.add(subAggregators[i]);
             }
         }
-        if (recordingWrapper != null) {
-            recordingWrapper.setDeferredCollector(deferredAggregations);
-            collectors.add(recordingWrapper);
+        if (deferringCollector != null) {
+            deferringCollector.setDeferredCollector(deferredAggregations);
+            collectors.add(deferringCollector);
         }
         collectableSubAggregators = MultiBucketCollector.wrap(collectors);
     }
 
-    public DeferringBucketCollector getDeferringCollector() {
-        // Default impl is a collector that selects the best buckets
-        // but an alternative defer policy may be based on best docs.
+    /**
+     * Get the deferring collector.
+     */
+    protected DeferringBucketCollector deferringCollector() {
+        return deferringCollector;
+    }
+
+    /**
+     * Build the {@link DeferringBucketCollector}. The default implementation
+     * selects the best buckets but some aggs want to do something else.
+     */
+    protected DeferringBucketCollector buildDeferringCollector() {
         return new BestBucketsDeferringCollector(topLevelQuery(), searcher(), descendsFromGlobalAggregator(parent()));
     }
 
@@ -92,8 +101,8 @@ public abstract class DeferableBucketAggregator extends BucketsAggregator {
 
     @Override
     protected final void prepareSubAggs(long[] bucketOrdsToCollect) throws IOException {
-        if (recordingWrapper != null) {
-            recordingWrapper.prepareSelectedBuckets(bucketOrdsToCollect);
+        if (deferringCollector != null) {
+            deferringCollector.prepareSelectedBuckets(bucketOrdsToCollect);
         }
     }
