+++ server/src/main/java/org/elasticsearch/search/aggregations/bucket/MergingBucketsDeferringCollector.java
@@ -21,11 +21,7 @@ package org.elasticsearch.search.aggregations.bucket;
 
 import org.apache.lucene.search.IndexSearcher;
 import org.apache.lucene.search.Query;
-import org.apache.lucene.util.packed.PackedInts;
-import org.apache.lucene.util.packed.PackedLongValues;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.function.LongUnaryOperator;
 
 /**
@@ -33,7 +29,9 @@ import java.util.function.LongUnaryOperator;
  * matches and then is able to replay a given subset of buckets. Exposes
  * mergeBuckets, which can be invoked by the aggregator when increasing the
  * rounding interval.
+ * @deprecated Use {@link BestBucketsDeferringCollector}
  */
+@Deprecated
 public class MergingBucketsDeferringCollector extends BestBucketsDeferringCollector {
     public MergingBucketsDeferringCollector(Query topLevelQuery, IndexSearcher searcher, boolean isGlobal) {
         super(topLevelQuery, searcher, isGlobal);
@@ -54,89 +52,10 @@ public class MergingBucketsDeferringCollector extends BestBucketsDeferringCollec
      *  This process rebuilds the ordinals and docDeltas according to the mergeMap, so it should
      *  not be called unless there are actually changes to be made, to avoid unnecessary work.
      *
-     * @deprecated use {@link mergeBuckets(LongUnaryOperator)}
+     * @deprecated use {@link BestBucketsDeferringCollector#rewriteBuckets(LongUnaryOperator)}
      */
     @Deprecated
     public void mergeBuckets(long[] mergeMap) {
-        mergeBuckets(bucket -> mergeMap[Math.toIntExact(bucket)]);
-    }
-
-    /**
-     * Merges/prunes the existing bucket ordinals and docDeltas according to the provided mergeMap.
-     *
-     * @param mergeMap a unary operator which maps a bucket's ordinal to the ordinal it should be merged with.
-     * If a bucket's ordinal is mapped to -1 then the bucket is removed entirely.
-     *
-     * This process rebuilds the ordinals and docDeltas according to the mergeMap, so it should
-     * not be called unless there are actually changes to be made, to avoid unnecessary work.
-     */
-    public void mergeBuckets(LongUnaryOperator mergeMap){
-        List<Entry> newEntries = new ArrayList<>(entries.size());
-        for (Entry sourceEntry : entries) {
-            PackedLongValues.Builder newBuckets = PackedLongValues.packedBuilder(PackedInts.DEFAULT);
-            PackedLongValues.Builder newDocDeltas = PackedLongValues.packedBuilder(PackedInts.DEFAULT);
-            PackedLongValues.Iterator docDeltasItr = sourceEntry.docDeltas.iterator();
-
-            long lastGoodDelta = 0;
-            for (PackedLongValues.Iterator itr = sourceEntry.buckets.iterator(); itr.hasNext();) {
-                long bucket = itr.next();
-                assert docDeltasItr.hasNext();
-                long delta = docDeltasItr.next();
-
-                // Only merge in the ordinal if it hasn't been "removed", signified with -1
-                long ordinal = mergeMap.applyAsLong(bucket);
-
-                if (ordinal != -1) {
-                    newBuckets.add(ordinal);
-                    newDocDeltas.add(delta + lastGoodDelta);
-                    lastGoodDelta = 0;
-                } else {
-                    // we are skipping this ordinal, which means we need to accumulate the
-                    // doc delta's since the last "good" delta
-                    lastGoodDelta += delta;
-                }
-            }
-            // Only create an entry if this segment has buckets after merging
-            if (newBuckets.size() > 0) {
-                assert newDocDeltas.size() > 0 : "docDeltas was empty but we had buckets";
-                newEntries.add(new Entry(sourceEntry.context, newDocDeltas.build(), newBuckets.build()));
-            }
-        }
-        entries = newEntries;
-
-        // if there are buckets that have been collected in the current segment
-        // we need to update the bucket ordinals there too
-        if (bucketsBuilder != null && bucketsBuilder.size() > 0) {
-            PackedLongValues currentBuckets = bucketsBuilder.build();
-            PackedLongValues.Builder newBuckets = PackedLongValues.packedBuilder(PackedInts.DEFAULT);
-            PackedLongValues.Builder newDocDeltas = PackedLongValues.packedBuilder(PackedInts.DEFAULT);
-
-            // The current segment's deltas aren't built yet, so build to a temp object
-            PackedLongValues currentDeltas = docDeltasBuilder.build();
-            PackedLongValues.Iterator docDeltasItr = currentDeltas.iterator();
-
-            long lastGoodDelta = 0;
-            for (PackedLongValues.Iterator itr = currentBuckets.iterator(); itr.hasNext();) {
-                long bucket = itr.next();
-                assert docDeltasItr.hasNext();
-                long delta = docDeltasItr.next();
-                long ordinal = mergeMap.applyAsLong(bucket);
-
-                // Only merge in the ordinal if it hasn't been "removed", signified with -1
-                if (ordinal != -1) {
-                    newBuckets.add(ordinal);
-                    newDocDeltas.add(delta + lastGoodDelta);
-                    lastGoodDelta = 0;
-                } else {
-                    // we are skipping this ordinal, which means we need to accumulate the
-                    // doc delta's since the last "good" delta.
-                    // The first is skipped because the original deltas are stored as offsets from first doc,
-                    // not offsets from 0
-                    lastGoodDelta += delta;
-                }
-            }
-            docDeltasBuilder = newDocDeltas;
-            bucketsBuilder = newBuckets;
-        }
+        rewriteBuckets(bucket -> mergeMap[Math.toIntExact(bucket)]);
     }
 }
